+ Deathwish 2 + All the resources till the end of dusts

+++ src/main/java/net/guwy/radiated/mechanics/hazard_items/IAsbestosItem.java
@@ -7,5 +7,5 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 
 public interface IAsbestosItem {
-    double asbestosVal(ItemStack itemStack);
+    int asbestosVal(ItemStack itemStack);
 }
